+++ NEWS.md
@@ -9,6 +9,17 @@ New features (CLI, Config):
 
 New features (Analysis):
 + Infer that `parent::someMethodReturningStaticType()` is a subtype of the current class, not just the parent class. (#4202)
++ Support phpdoc `@abstract` or `@phan-abstract` on non-abstract class constants, properties, and methods
+  to indicate that the intent is for non-abstract subclasses to override the definition. (#2278, #2285)
+  New issue types: `PhanCommentAbstractOnInheritedConstant`, `PhanCommentAbstractOnInheritedProperty`, `PhanCommentOverrideOnNonOverrideProperty`
+
+  For example, code using `static::SOME_CONST` or `static::$SOME_PROPERTY` or `$this->someMethod()`
+  may declare a placeholder `@abstract` constant/property/method,
+  and use this annotation to ensure that all non-abstract subclasses override the constant/property/method
+  (if using real abstract methods is not practical for a use case)
++ Warn about `@override` on properties that do not override an ancestor's property definition.
+  New issue type: `PhanCommentOverrideOnNonOverrideProperty`.
+  (Phan already warns for constants and methods)
 
 Bug fixes:
 + Properly analyze the right hand side for `$cond || throw ...;` (e.g. emit `PhanCompatibleThrowException`) (#4199)